JSS.Lib: Follow the spec a bit closer during parsing
We now "follow the steps" in the spec when we call Parse (however we
don't follow them strictly, for example our use of C# exceptions).

This also allows for a vm to be provided during parsing which allows
for the same vm to be used inside the CLI. This means that variables
(and other definitions) will be saved across executed lines.