chore: Implement an implicit operator for creating normal completions
Whenever the spec tells us to return a value for a function that returns
a completion, it wants us to wrap that value in a normal completion.

Before we had to manually make a normal completion using the
NormalCompletion static functions.

However, it is much better to have an implicit operator that will call
NormalCompletion for us and makes the code a lot more readable.