chore: Implement an "AbruptOr<T>" class
The AbruptOr class models an abrupt completion or a raw value.

This is used for spec functions or implementation defined functions that
return one type of value or an abrupt completion.

For example, ToUint32 either returns an abrupt completion or number.
Specifically, we want the return type of that speicific function to
return a uint.