feat: Return the actual property inside OrdinaryGetOwnProperty
The spec wants us to return a property descriptor not the actual value.
This allows support for undefined being resolved to work.

Before, when getting an undefined property value from object, we'd check if a
value was undefined and if it was we returned that the property doesn't
exist.

Now we can have properties that are defined but have a value of
undefined.