feat: Instantiate intrinsics per realm instead of using singletons
We now instantiate intrinsics (such as Object.prototype) per realm. This
is done in the CreateIntrinsics spec function.

We deviate from the spec slightly, as we don't use a record and just
have an instance variable per intrinsic on the realm.

This also means if there are two realms, someone changing an intrinsic
is not visible on the other realm.