JSS-8 Change Binding to a reference typed object
We were getting the binding from a dictionary and because Binding was a
value type, it would be copied locally. When we changed the value in
PutValue, it would only be copied to the local Binding instead of the
Binding in the dictionary.

Now, Binding is a reference type so the change is visible to the
dictionary.