JSS-13 Support strict (in)equality of Objects using their identity
We now support strict (in)equality of two Objects. If they have the same
reference in C# that means they have the same "identity". There might be
edge cases in spec that we don't know about, but for now C# reference
equality is the same as identity equality.

Example Code:
```js
var a = {};
var b = {};
var c = a;
a === b; // Returns false, used to crash
a === b; // Returns true, used to crash
```