JSS-37 Implement creation of a VM with required harness files executed
We now have a function for creating a VM that is isolated from any other
created before and has the required harness files evaluated on it.

This is required by INTERPRETING.md and allows for test case files to be
executed on it.

Currently, we execute every harness file that we can currently parse and
execute, however, in the future we should only execute the harness files
actually required by a test file.