JSS-38 Implement execution of test-262 test cases
When StartRunner is called on a Test262Runner, we enumerate every
JavaScript file in the test directory and evaluate them in individual
realms and vms as required by test-262.

We currently have limited logging and only have success and failure
instead of more detailed failure reasons (e.g. a parsing error).