JSS-40 Handle metadata based on executing harness files
We now respect the "raw" flag (apart from strict-mode) and the
"includes" list.

The raw flag signifies that the test case should not be
modified in anyway and should not have harness files executed.

The includes list is a list of additional harness files to be executed
along side sta.js and assert.js.