JSS-54 Use a copy of DataProperties to enumerate over in for-in loops
If a property was defined on an object being enumerated over in a for-in
loop, we would throw as we modified a collection that we were doing a
foreach loop on.

We now copy the data property dictionary instead of using the dictionary
directly. This prevents the collection from being modified while
evaluating the for-in loop.