JSS-63 Implement and maintain a "strictness stack" during AST execution
We now have a "strictness stack" that is maintained on the realm's
associated VM.

Whenever we start execution of a script we push the strictness of the
script and pop it after the script has finished execution.

Whenever we call a function we push the strictness of the function and
pop is afrer the function has finished execution. However, for
functions, if the surrounding code is already strict, we keep that
strictness.