JSS-63 Run test cases with correct strictness in test-262 runner
We now run test-262 test cases with the correct strictness according to
their metadata flags.

Normally, test cases will both be ran in strict and non-strict mode.

However, if the "raw" is set or the "onlyStrict" flag is not set then
we run in non-strict mode.

If the "raw" and "noStrict" flags are not set then we run in strict mode.