JSS-65 Implement automatic semi-colon rules 1 and 2
We now handle the rules 1 and 2 for automatic semicolon insertion. We
insertion parse semi-colons as being there if there is a line
terminator, a }, the last ) of a do-while loop or an end of file is
found.

Example Code:
```js
// let a let a would throw if ran as-is.
let a
let b // Works even if there are no semi-colons
```