JSS-66 Implement labelled evaluate for breakable statements
We now properly handle break completions inside of loops. We were
skipping the labelled evaluation of breakable statements as required in
the spec.

However, every breakable statement is now an IBreakableStatement that is
to be evalauted in LabelledEvaluation.

Example Code:
```js
for (;;) { break; } // Now returns undefined as a normal completion
```